trac dependency
fixes trac dependency for broken setuptools install.  setuptools doesn't
glean the revision of an svn checkout properly, causing it to not be
appended to the trac distribution version.  hgplugin requires a certain
revision of trac, and since the revision number isn't present, hgplugin
refuses to load.  this could cause version mismatches and odd behavior
if trac isn't upgraded with trac.